refactor(local-ci): extract footprint.py for disk-usage accounting
R2-1 fifth cut. Move 4 helpers (format_size_bytes, path_size_bytes,
local_ci_state_footprint, describe_path_for_cleanup) out of
local_ci.py into a sibling footprint.py module.

- local_ci.py: 10,961 → 10,907 lines (-54 net).
- footprint.py: 91 lines.
- footprint depends on state_paths for the bundles/prepared/logs/
  results/cloud-runs directory accessors — clean cross-module seam,
  same pattern as io_utils.

Tests: 213/213 local_ci tests pass.

Running total for Phase 4 R2-1: 5 modules extracted (state_paths,
normalize, git_helpers, io_utils, footprint). local_ci.py shrunk
11,285 → 10,907 lines (-378, -3.4%). Follow-up cuts will target the
queue persistence, bundle/sync transport, and recovery cluster.
